Elektor-style symbols
These symbols are designed to be used with the Altium database library system where all the 'pedigree' information is stored in the database and the symbols from this file are being referenced in the database. This makes re-use of common symbols much easier and you don't need to buy and install a Altium Vault package. 

Send me an e-mail if you have questions.

- Richard